requirements.txt
beautifulsoup4~=4.9.3
PyMySQL~=0.10.1
python-dotenv~=0.15.0